src/requirements.txt
# django-extensions
django-extensions==3.1.3
# postgresql
psycopg2-binary==2.8.6

# ipython
ipython==7.19.0
ipython-genutils==0.2.0
# pyyaml
PyYAML==5.4.1
# django restframework
djangorestframework==3.12.2
# swagger
drf-yasg==1.20.0
asgiref==3.4.1
Django==3.2.8
django-environ==0.4.5
pytz==2021.3
sqlparse==0.4.2
typing-extensions==3.10.0.2
# celery
celery==5.1.2
django-celery-results==2.2.0
